ubuf_pic: prefer size_t array indicies in ubuf_pic_blit_alpha10

+++ lib/upipe/ubuf_pic.c
@@ -258,11 +258,11 @@ int ubuf_pic_blit_alpha10(struct ubuf *dest, struct ubuf *src,
             return err;
         }
 
-        int plane_hsize = extract_hsize / src_hsub / src_macropixel *
+        size_t plane_hsize = extract_hsize / src_hsub / src_macropixel *
                           src_macropixel_size / 2;
-        int plane_vsize = extract_vsize / src_vsub;
+        size_t plane_vsize = extract_vsize / src_vsub;
 
-        for (int i = 0; i < plane_vsize; i++) {
+        for (size_t i = 0; i < plane_vsize; i++) {
             uint16_t *real_dst = (uint16_t *)dest_buffer;
             const uint16_t *real_src = (const uint16_t *)src_buffer;
             const uint16_t *real_alpha = (const uint16_t *)(alpha_plane + alpha_stride * i * src_vsub);
@@ -270,7 +270,7 @@ int ubuf_pic_blit_alpha10(struct ubuf *dest, struct ubuf *src,
             if ((!alpha_plane && alpha == 0x3ff) || threshold == 0) {
                 memcpy(dest_buffer, src_buffer, plane_hsize*2);
             } else if (!alpha_plane) {
-                for (int j = 0; j < plane_hsize; j++) {
+                for (size_t j = 0; j < plane_hsize; j++) {
                     real_dst[j] = (real_dst[j] * (0x3ff - alpha) + real_src[j] * alpha) / 0x3ff;
                 }
             } else if (threshold != 0x3ff) {
@@ -278,12 +278,12 @@ int ubuf_pic_blit_alpha10(struct ubuf *dest, struct ubuf *src,
                  * if alpha is over the threshold, we use the subpicture pixel.
                  */
                 if (alpha == 0x3ff) {
-                    for (int j = 0; j < plane_hsize; j++) {
+                    for (size_t j = 0; j < plane_hsize; j++) {
                         const uint16_t a = real_alpha[j * src_hsub];
                         if (a > threshold) real_dst[j] = real_src[j];
                     }
                 } else {
-                    for (int j = 0; j < plane_hsize; j++) {
+                    for (size_t j = 0; j < plane_hsize; j++) {
                         const uint16_t a = real_alpha[j * src_hsub] * alpha / 0x3ff;
                         if (a > threshold) real_dst[j] = real_src[j];
                     }
@@ -291,12 +291,12 @@ int ubuf_pic_blit_alpha10(struct ubuf *dest, struct ubuf *src,
             } else {
                 /* smooth and slow blending */
                 if (alpha == 0x3ff) {
-                    for (int j = 0; j < plane_hsize; j++) {
+                    for (size_t j = 0; j < plane_hsize; j++) {
                         const uint16_t a = real_alpha[j * src_hsub];
                         real_dst[j] = (real_dst[j] * (0x3ff - a) + real_src[j] * a) / 0x3ff;
                     }
                 } else {
-                    for (int j = 0; j < plane_hsize; j++) {
+                    for (size_t j = 0; j < plane_hsize; j++) {
                         const uint16_t a = real_alpha[j * src_hsub] * alpha / 0x3ff;
                         real_dst[j] = (real_dst[j] * (0x3ff - a) + real_src[j] * a) / 0x3ff;
                     }
